% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{create_branch}
\alias{create_branch}
\title{Creates a new Branch for an Amplify App}
\usage{
create_branch(
  appId = "string",
  branchName = "string",
  description = "string",
  stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL",
  framework = "string",
  enableNotification = TRUE|FALSE,
  enableAutoBuild = TRUE|FALSE,
  environmentVariables = list(
    "string"
  ),
  basicAuthCredentials = "string",
  enableBasicAuth = TRUE|FALSE,
  tags = list(
    "string"
  ),
  buildSpec = "string",
  ttl = "string"
)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch.}

\item{description}{Description for the branch.}

\item{stage}{Stage for the branch.}

\item{framework}{Framework for the branch.}

\item{enableNotification}{Enables notifications for the branch.}

\item{enableAutoBuild}{Enables auto building for the branch.}

\item{environmentVariables}{Environment Variables for the branch.}

\item{basicAuthCredentials}{Basic Authorization credentials for the branch.}

\item{enableBasicAuth}{Enables Basic Auth for the branch.}

\item{tags}{Tag for the branch.}

\item{buildSpec}{BuildSpec for the branch.}

\item{ttl}{The content TTL for the website in seconds.}
}
\description{
Creates a new Branch for an Amplify App.
}
\examples{

}
