% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{create_domain_association}
\alias{create_domain_association}
\title{Create a new DomainAssociation on an App}
\usage{
create_domain_association(
  appId = "string",
  domainName = "string",
  enableAutoSubDomain = TRUE|FALSE,
  subDomainSettings = list(
    list(
      prefix = "string",
      branchName = "string"
    )
  )
)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{domainName}{[required] Domain name for the Domain Association.}

\item{enableAutoSubDomain}{Enables automated creation of Subdomains for branches.}

\item{subDomainSettings}{[required] Setting structure for the Subdomain.}
}
\description{
Create a new DomainAssociation on an App
}
\examples{

}
