% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{delete_app}
\alias{delete_app}
\title{Delete an existing Amplify App by appId}
\usage{
delete_app(appId)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}
}
\description{
Delete an existing Amplify App by appId.
}
\section{Accepted Parameters}{
\preformatted{delete_app(
  appId = "string"
)
}
}

