% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{delete_branch}
\alias{delete_branch}
\title{Deletes a branch for an Amplify App}
\usage{
delete_branch(
  appId = "string",
  branchName = "string"
)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch.}
}
\description{
Deletes a branch for an Amplify App.
}
\examples{

}
