% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{delete_domain_association}
\alias{delete_domain_association}
\title{Deletes a DomainAssociation}
\usage{
delete_domain_association(
  appId = "string",
  domainName = "string"
)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{domainName}{[required] Name of the domain.}
}
\description{
Deletes a DomainAssociation.
}
\examples{

}
