% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{delete_job}
\alias{delete_job}
\title{Delete a job, for an Amplify branch, part of Amplify App}
\usage{
delete_job(appId, branchName, jobId)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch, for the Job.}

\item{jobId}{[required] Unique Id for the Job.}
}
\description{
Delete a job, for an Amplify branch, part of Amplify App.
}
\section{Accepted Parameters}{
\preformatted{delete_job(
  appId = "string",
  branchName = "string",
  jobId = "string"
)
}
}

