% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{get_app}
\alias{get_app}
\title{Retrieves an existing Amplify App by appId}
\usage{
get_app(appId)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}
}
\description{
Retrieves an existing Amplify App by appId.
}
\section{Accepted Parameters}{
\preformatted{get_app(
  appId = "string"
)
}
}

