% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{get_branch}
\alias{get_branch}
\title{Retrieves a branch for an Amplify App}
\usage{
get_branch(appId, branchName)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch.}
}
\description{
Retrieves a branch for an Amplify App.
}
\section{Accepted Parameters}{
\preformatted{get_branch(
  appId = "string",
  branchName = "string"
)
}
}

