% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{list_apps}
\alias{list_apps}
\title{Lists existing Amplify Apps}
\usage{
list_apps(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{Pagination token. If non-null pagination token is returned in a result, then pass its value in another request to fetch more entries.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
Lists existing Amplify Apps.
}
\section{Accepted Parameters}{
\preformatted{list_apps(
  nextToken = "string",
  maxResults = 123
)
}
}

