% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{list_domain_associations}
\alias{list_domain_associations}
\title{List domains with an app}
\usage{
list_domain_associations(
  appId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{nextToken}{Pagination token. Set to null to start listing Apps from start. If non-null pagination token is returned in a result, then pass its value in here to list more projects.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
List domains with an app
}
\examples{

}
