% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_package.R
\docType{package}
\name{paws.amplify-package}
\alias{paws.amplify}
\alias{paws.amplify-package}
\title{paws.amplify: AWS Amplify}
\description{
Amplify is a fully managed continuous deployment and hosting
service for modern web apps.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
