% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{start_job}
\alias{start_job}
\title{Starts a new job for a branch, part of an Amplify App}
\usage{
start_job(appId, branchName, jobId = NULL, jobType, jobReason = NULL,
  commitId = NULL, commitMessage = NULL, commitTime = NULL)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch, for the Job.}

\item{jobId}{Unique Id for the Job.}

\item{jobType}{[required] Type for the Job.}

\item{jobReason}{Reason for the Job.}

\item{commitId}{Commit Id from 3rd party repository provider for the Job.}

\item{commitMessage}{Commit message from 3rd party repository provider for the Job.}

\item{commitTime}{Commit date / time for the Job.}
}
\description{
Starts a new job for a branch, part of an Amplify App.
}
\section{Accepted Parameters}{
\preformatted{start_job(
  appId = "string",
  branchName = "string",
  jobId = "string",
  jobType = "RELEASE"|"RETRY"|"WEB_HOOK",
  jobReason = "string",
  commitId = "string",
  commitMessage = "string",
  commitTime = as.POSIXct("2015-01-01")
)
}
}

