% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{update_branch}
\alias{update_branch}
\title{Updates a branch for an Amplify App}
\usage{
update_branch(appId, branchName, description = NULL, framework = NULL,
  stage = NULL, enableNotification = NULL, enableAutoBuild = NULL,
  environmentVariables = NULL, basicAuthCredentials = NULL,
  enableBasicAuth = NULL, buildSpec = NULL, ttl = NULL)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{branchName}{[required] Name for the branch.}

\item{description}{Description for the branch.}

\item{framework}{Framework for the branch.}

\item{stage}{Stage for the branch.}

\item{enableNotification}{Enables notifications for the branch.}

\item{enableAutoBuild}{Enables auto building for the branch.}

\item{environmentVariables}{Environment Variables for the branch.}

\item{basicAuthCredentials}{Basic Authorization credentials for the branch.}

\item{enableBasicAuth}{Enables Basic Auth for the branch.}

\item{buildSpec}{BuildSpec for the branch.}

\item{ttl}{The content TTL for the website in seconds.}
}
\description{
Updates a branch for an Amplify App.
}
\section{Accepted Parameters}{
\preformatted{update_branch(
  appId = "string",
  branchName = "string",
  description = "string",
  framework = "string",
  stage = "PRODUCTION"|"BETA"|"DEVELOPMENT"|"EXPERIMENTAL",
  enableNotification = TRUE|FALSE,
  enableAutoBuild = TRUE|FALSE,
  environmentVariables = list(
    "string"
  ),
  basicAuthCredentials = "string",
  enableBasicAuth = TRUE|FALSE,
  buildSpec = "string",
  ttl = "string"
)
}
}

