% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{update_domain_association}
\alias{update_domain_association}
\title{Create a new DomainAssociation on an App}
\usage{
update_domain_association(appId, domainName, enableAutoSubDomain = NULL,
  subDomainSettings)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{domainName}{[required] Name of the domain.}

\item{enableAutoSubDomain}{Enables automated creation of Subdomains for branches.}

\item{subDomainSettings}{[required] Setting structure for the Subdomain.}
}
\description{
Create a new DomainAssociation on an App
}
\section{Accepted Parameters}{
\preformatted{update_domain_association(
  appId = "string",
  domainName = "string",
  enableAutoSubDomain = TRUE|FALSE,
  subDomainSettings = list(
    list(
      prefix = "string",
      branchName = "string"
    )
  )
)
}
}

