% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.amplify_operations.R
\name{get_domain_association}
\alias{get_domain_association}
\title{Retrieves domain info that corresponds to an appId and domainName}
\usage{
get_domain_association(appId, domainName)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{domainName}{[required] Name of the domain.}
}
\description{
Retrieves domain info that corresponds to an appId and domainName.
}
\section{Accepted Parameters}{
\preformatted{get_domain_association(
  appId = "string",
  domainName = "string"
)
}
}

